Game functions grouped and moved to custom package
